+++ AlchemistCode/Cards/Hero/Overspray.cs
@@ -1,0 +1,23 @@
+using Alchemist.AlchemistCode.Powers;
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models.Powers;
+
+namespace Alchemist.AlchemistCode.Cards.Hero;
+
+[CardTheme(CardTheme.Poison)]
+public class Overspray : AlchemistHeroCard
+{
+    public Overspray() : base(2, CardType.Power, CardRarity.Rare, TargetType.Self)
+    {
+        WithCostUpgradeBy(-1);
+        WithTip(typeof(PoisonPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await PowerCmd.Apply<OversprayPower>(choiceContext, Owner.Creature, 1, Owner.Creature, this);
+    }
+}