+++ AlchemistCode/Cards/Hero/PondScum.cs
@@ -1,0 +1,26 @@
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models.Powers;
+using MegaCrit.Sts2.Core.ValueProps;
+
+namespace Alchemist.AlchemistCode.Cards.Hero;
+
+[CardTheme(CardTheme.Poison)]
+public class PondScum : AlchemistHeroCard
+{
+    private const int PerEnemy = 2;
+
+    protected internal override bool PlaysCastAnimation => false;
+
+    public PondScum() : base(1, CardType.Skill, CardRarity.Common, TargetType.Self)
+    {
+        WithCalculatedBlock(5, static (card, _) => PoisonedEnemies(card) * PerEnemy, ValueProp.Move, 2, 0);
+        WithTip(typeof(PoisonPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await CommonActions.CardBlock(this, play);
+    }
+}